convert tests to use tasty
The purpose of this commit is to simplify the test code and get this
package back into stackage nightly be removing it's dependencies on
deprecated packages. The deprecated package replaced by this commit is
`testpack`. Addresses problems in issue #12.

I believe that this also simplifies the test code somewhat, but someone
more familiar with the tests should probably take a look to make sure I
didn't mess up some tests.

stack build (with tests enabled) and stack exec runtests worked on
against stackage nightly.